+++ src/Traits/ManagesItemsTrait.php
@@ -245,12 +245,14 @@ trait ManagesItemsTrait
     }
 
     /**
-     * Returns json serialized representation of array of items
+     * Get the collection of items as JSON.
+     *
+     * @param  int  $options
      * @return string
      */
-    public function toJson()
+    public function toJson($options = 0)
     {
-        return json_encode($this->getAll());
+        return json_encode($this->getAll(), $options);
     }
 
     /**